refactor(Member): Member 분리 및 테스트 수정
