+++ src/test/java/domain/MemberTest.java
@@ -1,8 +1,8 @@
 package domain;
 
-import domain.vo.RoundResult;
 import domain.card.Card;
 import domain.member.Dealer;
+import domain.member.Members;
 import domain.member.Player;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
@@ -10,10 +10,17 @@ import org.junit.jupiter.api.Test;
 
 public class MemberTest {
 
+    @DisplayName("딜러이면 true를 반환한다.")
+    @Test
+    void roleTest_isDealer_returnTrue() {
+        Members members = new Members();
+        Assertions.assertTrue(members.isDealer("딜러"));
+    }
+
     @DisplayName("플레이어가 이기면 RoundResult를 반환한다.")
     @Test
     void winnerTest_isPlayerThenDecide_returnMember() {
-        Dealer dealer = new Dealer();
+        Dealer dealer = new Dealer("브라운");
         Player player = new Player("브리");
 
         dealer.receiveCard(new Card("2", "하트"));
@@ -21,6 +28,6 @@ public class MemberTest {
         player.receiveCard(new Card("4","다이아몬드"));
         player.receiveCard(new Card("2", "스페이드"));
 
-        Assertions.assertEquals(RoundResult.WIN, RoundResult.judgeAgainst(dealer.handValue(), player.handValue()));
+        Assertions.assertEquals(player.judgeAgainst(dealer), RoundResult.WIN);
     }
 }
